app/requirements.txt
flask
gunicorn
tensorflow-cpu
transformers